Add HTML5 video support
Requires the underlying GNU social instance to have the VideoThumbnails
plugin and (for that) 'libav-tools' (with 'avconv') installed. Support
for ffmpeg or maybe mplayer as a fallback is considered but not planned.

If the underlying instance does _not_ support video thumbnailing there
will simply be none to fetch and none will be listed.

One bug remains and that is when the video is clicked (on the play button)
the notice closes again. I am not sure how to abort that properly.