No need to check logged in user
It's always the notice's author we want to check this for anyway.

I kept getting errors like, which are now fixed:
   Trying to get property of non-object in /srv/www/vhosts/se/umeahackerspace/social/%/htdocs/local/plugins/Qvitter/QvitterPlugin.php on line 594

I guess it might also screw up (somehow?) if some future awesome
notice handling thing decided to run the StartNoticeDistribute event
when the session is not currently used by the notice's author. But
that's a bit far-fetched.