Only enable qvitter for javascript users
Also makes it less problematic that Qvitter doesn't provide <head/>
metadata everywhere.

The first time a non-logged in user loads the website, javascript will
run (or not) that sets a cookie to enable qvitter and reloads the page.

I also made it so that when Qvitter is enabled, it just reloads the same
page and doesn't go to your home timeline.