MediaFile::fromFilehandle on profile avatar upload
This is mainly because I in GNU social want to change the constructor of
MediaFile and thus use fromFilehandle in Qvitter, so we don't have to sync
the changes as much.

This also will make sure that uploaded files can be reused if they match a
SHA256 hash of the avatar we want to set. Parts of my rewriting also provide
a pretty good example of how we can reuse existing File entries as Avatar sources.