FileNotFoundException won't be thrown if MediaFile works as it should
In newer GNU social MediaFile will make sure the local file exists, if
we're trying to upload something that matches by hash in the File table.
For the current 'master' when I'm pushing this that won't be handled
correctly, but then again it would only be a problem if something has
been f'd already and you're missing files from the filesystem that are
listed in the File table.